refactor 포인트 도메인 수정 #13

+++ src/main/java/kr/hhplus/be/server/interfaces/product/ProductResponse.java
@@ -4,11 +4,13 @@ package kr.hhplus.be.server.interfaces.product;
 
 import kr.hhplus.be.server.application.product.ProductResult;
 
+import java.math.BigDecimal;
+
 public record ProductResponse(
         Long id,
         String name,
-        int price,
-        int stock
+        BigDecimal price,
+        Long stock
 ) {
     public static ProductResponse fromResult(ProductResult result) {
         return new ProductResponse(
